Refactor MoveValidator for better readability
